requirements.txt
aiohttp==3.4.0
yarl<1.2
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]
pygithub
Flask
requests
